requirements.txt
appdirs==1.4.4
asttokens==2.4.1
beautifulsoup4==4.12.3
borb==2.1.21
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cmake==3.29.0.1
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cryptography==42.0.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distlib==0.3.8
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.15.4
fonttools==4.48.1
frozendict==2.4.0
fsspec==2024.6.1
html5lib==1.1
idna==3.6
intel-openmp==2021.4.0
ipykernel==6.29.2
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
keyboard==0.13.5
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.6
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
parso==0.8.3
peewee==3.17.1
pickle-mixin==1.0.2
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyparsing==3.1.2
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
python-barcode==0.15.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.1
pywin32==306
pyzmq==25.1.2
qrcode==7.4.2
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
setuptools==69.0.3
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.1
tbb==2021.13.0
threadpoolctl==3.2.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.36
